Changing calculateElementWidth for old IE
In IE7 and IE8 (and maybe older, and maybe other browsers)
el.css('marginLeft') will return 'auto' instead of 0px. So I'm
checking to see if that value was returned, because parseInt will
return a NaN if we give it 'auto'. Also removing the margin in our
test fixture and updating tests.